#26 Replace :x: failed build emoji with :cross_mark:
While the :x: emoji used to be x in a red box in Zulip 2.x, it was
replaced with green box in Zulip 3.x and was not suitable to
communicate failed builds anymore.
The borderless :cross_mark: replacement will also align better
with the :check_mark: used for successful builds.